requirements.txt
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
Pygments==2.2.0
pynput==1.3.10
pyobjc-core==4.2.2
pyobjc-framework-Cocoa==4.2.2
pyobjc-framework-Quartz==4.2.2
six==1.11.0
virtualenv==15.2.0
